rest/requirements.txt
# For the DB
psycopg2-binary

# For flask
flask
flask-cors
requests